+++ .github/workflows/Consolidation-without-Anchors.yml
@@ -30,9 +30,7 @@ jobs:
 
       - name: Check HTML links
         continue-on-error: true
-        run: |
-          sudo apt-get install w3c-checklink
-          w3c-checklink -r public > htmltest.log
+        run: bin/htmltest -c .htmltest.yml -s 2>&1 | tee htmltest.log
 
       - name: HTML test results
         uses: actions/upload-artifact@v2
