+++ .profanity_ignore.yml
@@ -1,0 +1,12 @@
+XX
+-kill
+kill
+KVM
+#IO_L3N_T0_DQS_AD1N_35
+172.X.X.X
+naked
+(x=x
+**VM
+Kill
+slave
+Slave
